add routes folder and restructure code
