+++ back-end/utils/database.js
@@ -1,12 +1,9 @@
 const mongoose = require("mongoose");
 
 const env = process.env.NODE_ENV || 'development';
+const dbURI = env === "test" ? 'mongodb://localhost/foodWithFriends' : process.env.DB_CONNECTION;
 
-if (env === 'test') {
-  process.env.DB_CONNECTION = 'mongodb://localhost/foodWithFriends'
-}
-
-mongoose.connect(process.env.DB_CONNECTION, {
+mongoose.connect(dbURI, {
     useNewUrlParser: true, 
     useUnifiedTopology: true
   }, () => {
